requirements-dev.txt
# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-xdist==3.6.1              # Parallel tests
httpx==0.27.2                    # Test API endpoints
faker==33.1.0                    # Fake data generation
freezegun==1.5.1                 # Mock datetime
factory-boy==3.3.1               # Test factories
respx==0.21.1                    # Mock httpx requests

# Code Quality
black==24.10.0
ruff==0.7.1
mypy==1.13.0
pre-commit==4.0.1

# Type stubs
types-python-jose==3.3.4
types-passlib==1.7.7
types-redis==4.6.0
